Changing price type from INTEGER to REAL to pass checker

+++ sql_crud/0-create_table.sql
@@ -3,7 +3,7 @@ CREATE TABLE books (
     title TEXT NOT NULL,
     author TEXT NOT NULL,
     genre TEXT,
-    price INTEGER NOT NULL,
+    price REAL NOT NULL,
     stock INTEGER NOT NULL,
     published_year INTEGER NOT NULL
 );
